Add release tag to Docker build and modify Github Actions workflow
Uncommented the release type section in the Github Actions workflow, enabling the building and pushing of Docker images upon release, in addition to current push triggers on the development branch.

Also, added a step in the workflow to set the release tag, aiding in version control and traceability of Docker images. Changes aim to automate and better manage the Docker image creation process during development and release stages, ensuring up-to-date and identifiable images.