fix: Update mask expansion in ImagePreprocessor
This commit refactors code in segmentation_utils/ImagePreprocessor.py to improve compatibility. Specifically, it adds a condition to check the dimensionality of the mask before expanding it with np.expand_dims, thereby guarding against unnecessary expansion and potential errors. This refurbishment not only improves code maintainability, but also enhances performance of the image and mask predifining augmentation functions.