Add dynamic backend support and improve image preprocessing
This commit introduces the capability of dynamic backend switching between Tensorflow and Pytorch. This enhances the project's flexibility and customization. Files for implementing both backends are added under utilities. In the ImagePreprocessor.py, batch_images and batch_masks are converted to float32 type for better compatibility. A README.md has been updated with guidance on how to switch backends. The commit also improves test coverage and includes minor adjustments in tests/README.md to improve readability.