Enable Docker Build and Push on release
This commit changes the `.github/workflows/docker_landseg.yml` file to enable the Docker Build and Push action specifically on release events rather than on push to the 'development' branch. This ensures that Docker builds and pushes are only made when a release is created, aligning the action more closely to the release workflow.