Add 'curl' to Dockerfile dependencies
The 'curl' package was added to the Dockerfile in order to provide essential command-line utilities for requesting and handling data from servers. It allows us to download or upload data while making use of a variety of protocols, which can be crucial in the context of Docker and its automation capability.